Put all decorator classes in one file
